починил баг с ресурсами и энергией

+++ src/reducer/resourceSlice.jsx
@@ -58,17 +58,29 @@ export const addResourceToInventory = createAsyncThunk(
 export const removeResource = createAsyncThunk(
   "remove/Resource",
   async (resourceName, thunkAPI) => {
-    try {
-      const res = await fetch(`http://localhost:4000/resources/${resourceName}`, {
-        method: "DELETE",
-      });
+      try {
+          const token = localStorage.getItem("token");
 
-      if (res.ok) {
-        return resourceName;
+          if (!token) {
+              throw new Error("Token is not defined");
+          }
+
+          const response = await fetch(`http://localhost:4000/resources/eat/Ягоды`, {
+              method: 'DELETE',
+              headers: {
+                  'Content-Type': 'application/json',
+                  Authorization: `Bearer ${localStorage.getItem('token')}`,
+              },
+          });
+
+          if (response.ok) {
+              return resourceName;
+          } else if (response.status === 404) {
+              return thunkAPI.rejectWithValue("Resource not found");
+          }
+      } catch (error) {
+          return thunkAPI.rejectWithValue(error);
       }
-    } catch (error) {
-      return thunkAPI.rejectWithValue(error);
-    }
   }
 );
